ref: able to write from nvbar action button
